+++ internal/bootstrap/setup.go
@@ -5,6 +5,8 @@ import (
 	"personal-finance/internal/bootstrap/invoice"
 	"personal-finance/internal/bootstrap/movement"
 	"personal-finance/internal/bootstrap/registry"
+	"personal-finance/internal/bootstrap/transfer"
+	"personal-finance/internal/bootstrap/userpreferences"
 
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
@@ -16,4 +18,6 @@ func SetupCleanArchComponents(r *gin.Engine, db *gorm.DB) {
 	movement.Setup(r, reg)
 	creditcard.Setup(r, reg)
 	invoice.Setup(r, reg)
+	transfer.Setup(r, reg)
+	userpreferences.Setup(r, reg)
 }